% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_project_sync_off.R
\name{dropbox_project_sync_off}
\alias{dropbox_project_sync_off}
\title{Make Dropbox ignore .Rproj folders}
\usage{
dropbox_project_sync_off()
}
\description{
Make Dropbox ignore .Rproj folders
}
