% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpcomp.r
\name{ebpcomp}
\alias{ebpcomp}
\title{ebpcomp}
\usage{
ebpcomp(x, qref = c(0.5, 0.25, 0.75), probs = c(0.05, 0.125, 0.25, 0.375))
}
\arguments{
\item{x}{a numeric variable}

\item{qref}{quantiles for major corners}

\item{probs}{quantiles for minor corners}
}
\value{
list with elements \code{segments}, \code{lines}, \code{points}, \code{points2}
}
\description{
Computation of Coordinates of Extended Box Plots Elements
}
\details{
For an extended box plots computes all the elements needed for plotting it.  This is typically used when adding to a \code{ggplot2} plot.
}
\examples{
ebpcomp(1:1000)
}
\author{
Frank Harrell
}
