% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fdebug.r
\name{Fdebug}
\alias{Fdebug}
\title{Debug Printing Function Generator}
\usage{
Fdebug(opt)
}
\arguments{
\item{opt}{an unquoted option name}
}
\value{
a function
}
\description{
Takes the name of a system \code{option(opt=)} and checks to see if option \code{opt} is
set to \code{TRUE}, taking its default value to be \code{FALSE}.  If \code{TRUE}, a function is
created that calls \code{\link[=prn]{prn()}} to print an object with the object's name in the
description along with the option name and the name of the function within which
the generated function was called, if any.  If option \code{opt} is not set, a dummy function
is generated instead.
}
\examples{
dfun <- Fdebug(my_option_name)   # my_option_name not currently set
dfun
dfun(sqrt(2))
}
\author{
Fran Harrell
}
