% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abd.r
\name{getabd}
\alias{getabd}
\title{getabd}
\usage{
getabd(name = "", lowernames = FALSE, allow = "_")
}
\arguments{
\item{name}{name of dataset to fetch.  Omit to get a data table listing all available datasets.}

\item{lowernames}{set to \code{TRUE} to change variable names to lower case}

\item{allow}{set to \code{NULL} to convert underscores in variable names to periods}
}
\value{
data frame with attributes \code{label} and \code{url}
}
\description{
Data from The Analysis of Biological Data by Shitlock and Schluter
}
\details{
Fetches csv files for exercises in the book
}
\author{
Frank Harrell
}
