% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotlyr.r
\name{ggplotlyr}
\alias{ggplotlyr}
\title{ggplotlyr}
\usage{
ggplotlyr(ggobject, tooltip = "label", remove = "txt: ", ...)
}
\arguments{
\item{ggobject}{an object produced by \code{ggplot}}

\item{tooltip}{attribute specified to \code{ggplot} to hold hover text}

\item{remove}{extraneous text to remove from hover text.  Default is set to assume \code{tooltip='label'} and assumed the user specified \code{aes(..., label=txt)}.  If you instead specified \code{aes(..., label=myvar)} use \code{remove='myvar: '}.}

\item{...}{other arguments passed to \code{ggplotly}}
}
\value{
a \code{plotly} object
}
\description{
Render \code{plotly} Graphic from a \code{ggplot2} Object
}
\details{
Uses \code{plotly::ggplotly()} to render a \code{plotly} graphic with a specified tooltip attribute, removing extraneous text that \code{ggplotly} puts in hover text when \code{tooltip='label'}
}
\author{
Frank Harrell
}
