% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_l.R, R/head_m.R, R/head_s.R
\name{head_l}
\alias{head_l}
\alias{head_m}
\alias{head_s}
\title{head() with commonly used number of observations(100, 1000, 10000)}
\usage{
head_l(obj)

head_m(obj)

head_s(obj)
}
\arguments{
\item{obj}{data object}
}
\description{
head() with commonly used number of observations(100, 1000, 10000)

head() with commonly used number of observations(100, 1000, 10000)

head() with commonly used number of observations(100, 1000, 10000)
}
\section{Functions}{
\itemize{
\item \code{head_l()}: head(n = 10000)

\item \code{head_m()}: head(n = 1000)

}}
