% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMarkovOrd.r
\name{intMarkovOrd}
\alias{intMarkovOrd}
\title{Compute Intercepts for Proportional Odds Markov Model}
\usage{
intMarkovOrd(y, times, initial, absorb = NULL, intercepts, g, target, t, ...)
}
\arguments{
\item{y}{see \link{simMarkovOrd}}

\item{times}{see \link{simMarkovOrd}}

\item{initial}{see \link{simMarkovOrd}}

\item{absorb}{see \link{simMarkovOrd}}

\item{intercepts}{vector of initial guesses for the intercepts}

\item{g}{see \link{simMarkovOrd}}

\item{target}{vector of target state occupancy probabilities at time \code{t}}

\item{t}{target time}

\item{...}{optional arguments to pass to \code{\link[stats:nlm]{stats::nlm()}}}
}
\value{
vector of intercept values
}
\description{
Given a vector \code{intercepts} of initial guesses at the intercepts in a Markov proportional odds model, solves for the intercept vector that yields a set of occupancy probabilities at time \code{t} that equal a vector of target values.
}
\seealso{
\url{https://hbiostat.org/R/Hmisc/simMarkovOrd.html}
}
\author{
Frank Harrell
}
