% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latestFile.r
\name{latestFile}
\alias{latestFile}
\title{latestFile}
\usage{
latestFile(pattern, path = ".", verbose = TRUE)
}
\arguments{
\item{pattern}{a regular expression; see \code{\link[base:list.files]{base::list.files()}}}

\item{path}{full path, defaulting to current working directory}

\item{verbose}{set to \code{FALSE} to not report on total number of matching files}
}
\value{
the name of the last modified file
}
\description{
Find File With Latest Modification Time
}
\details{
Subject to matching on \code{pattern} finds the last modified file, and if \code{verbose} is \code{TRUE} reports on how many total files matched \code{pattern}.
}
\seealso{
\code{\link[base:list.files]{base::list.files()}}
}
\author{
Frank Harrell
}
