\name{legendfunctions}
\alias{legendfunctions}
\alias{Key}
\alias{sKey}
\alias{Key2}
\alias{Points}
\alias{Abline}
\alias{Curve}
\title{Legend Creation Functions}
\description{
  Wrapers to plot defined legend ploting functions
}
\usage{
Key(...)
Key2(...)
sKey(...)
Points(...)
Abline(...)
Curve(...)
}
\arguments{
  \item{\dots}{arguments to pass to wrapped functions}
}
