% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_id.R
\name{load_id}
\alias{load_id}
\title{Automate loading with appropriate id code -- mirror of save_id}
\usage{
load_id(file = "./data/", no, id = saveid)
}
\arguments{
\item{file}{directory into which I wish to load from}

\item{id}{identification of the code that I am using ("107", etc.)}

\item{nm}{name of the file}
}
\description{
Automate loading with appropriate id code -- mirror of save_id
}
