% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{load data with a given name
Credit to the original author of the function:
https://stackoverflow.com/a/25455968/10214507}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{directory and name of file}
}
\value{
Load a data to a given name
}
\description{
load data with a given name
Credit to the original author of the function:
https://stackoverflow.com/a/25455968/10214507
}
\examples{
dataA <- loadRData("./data/tempdata.Rda")
}
