% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCoincident.r
\name{nCoincident}
\alias{nCoincident}
\title{nCoincident}
\usage{
nCoincident(x, y, bins = 400)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{bins}{number of bins in both directions}
}
\value{
integer count
}
\description{
Number of Coincident Points
}
\details{
Computes the number of x,y pairs that are likely to be obscured in a regular scatterplot, in the sense of overlapping pairs after binning into \code{bins} x \code{bins} squares where \code{bins} defaults to 400.  \code{NA}s are removed first.
}
\examples{
nCoincident(c(1:5, 4:5), c(1:5, 4:5)/10)
}
\author{
Frank Harrell
}
