% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princmp.r
\name{plot.princmp}
\alias{plot.princmp}
\title{plot.princmp}
\usage{
\method{plot}{princmp}(x, offset = 0.8, col = 1, adj = 0, ylim = NULL, add = FALSE)
}
\arguments{
\item{x}{results of `princmp`}

\item{offset}{controls positioning of text labels for cumulative fraction of variance explained}

\item{col}{color of plotted text}

\item{adj}{angle for plotting text}

\item{ylim}{y-axis plotting limits, a 2-vector}

\item{add}{set to `TRUE` to add a line to an existing plot without drawing axes}
}
\value{
nothing
}
\description{
Scree Plot for princmp
}
\details{
Uses base graphics to plot the scree plot from a [princmp()] result, showing cumultive proportion of variance explained
}
\author{
Frank Harrell
}
