% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.r
\name{plotlySave}
\alias{plotlySave}
\title{Save a Plotly Graphic to PNG}
\usage{
plotlySave(x, ...)
}
\arguments{
\item{x}{a \code{plotly} graphics object or a named list of such objects.
The resulting \code{png} file will go in the file path given by the
\code{knitr} \code{fig.path} value, and have a base name equal to the current
\code{knitr} chunk name. If \code{x} is a list, a minus sign followed by the
chunk name are inserted before \code{.png}.}

\item{...}{additional arguments passed to \code{plotly::plotly_IMAGE}}
}
\value{
called for its side effect of writing one or more PNG files;
returns \code{invisible(NULL)}
}
\description{
\code{plotlySave} saves a plotly graphic with name \code{foo.png} where \code{foo} is
the name of the current chunk. You must have a free \code{plotly} account
from \code{plot.ly} to use this function, and you must have run
\code{Sys.setenv(plotly_username="your_plotly_username")} and
\code{Sys.setenv(plotly_api_key="your_api_key")}. The API key can be found
in one's profile settings. See
\url{http://stackoverflow.com/questions/33959635/exporting-png-files-from-plotly-in-r}.
}
\examples{
# ```{r chunkname}
# p <- plotly::plot_ly(...)
# plotlySave(p)   # creates fig.path/chunkname.png
# ```

}
