% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.r
\name{plotp}
\alias{plotp}
\title{Generic Method for Making Plotly Graphics}
\usage{
plotp(data, ...)
}
\arguments{
\item{data}{an object having a \code{plotp} method}

\item{...}{additional arguments passed to the specific \code{plotp}
method}
}
\value{
typically a plotly object, as produced by the dispatched method
}
\description{
\code{plotp} is a generic function that dispatches to \code{plotp}
methods to make \code{plotly} graphics.
}
