% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princmp.r
\name{print.princmp}
\alias{print.princmp}
\title{print.princmp}
\usage{
\method{print}{princmp}(x, scoef = TRUE, orig = TRUE)
}
\arguments{
\item{x}{results of \code{princmp}}

\item{scoef}{set to \code{FALSE} to not print coefficients (loadings) of standardized variables}

\item{orig}{set to \code{FALSE} to not show coefficients on the original scale}
}
\value{
nothing
}
\description{
Print Results of princmp
}
\details{
Simple print method for \code{\link[=princmp]{princmp()}}
}
\author{
Frank Harrell
}
