% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Measures.r
\name{R2Measures}
\alias{R2Measures}
\title{R2Measures}
\usage{
R2Measures(lr, p, n, ess = NULL, padj = 1, inclval = FALSE)
}
\arguments{
\item{lr}{likelihoood ratio chi-square statistic}

\item{p}{number of non-intercepts in the model that achieved \code{lr}}

\item{n}{raw number of observations}

\item{ess}{if a single number, is the effective sample size.  If a vector of numbers is assumed to be the frequency tabulation of all distinct values of the outcome variable, from which the effective sample size is computed.}

\item{padj}{set to 2 to use the classical adjusted R^2 penalty, 1 to subtract \code{p} from \code{lr}}

\item{inclval}{set to \code{TRUE} to R2 values in \code{html} and \code{latex} attributes, otherwise only the labels for the measures are returned}
}
\value{
named a vector of R2 measures, with two attributes: \code{html} contains the character string results formatted for html, and \code{latex} contains a character string vector of results formatted for LaTeX.  If \code{inclval=TRUE} these representations contain also the numeric results rounded to 3 decimal places.  The notation for results is R^2 sub \verb{p, n} where the \code{p} component is empty for unadjusted estimtes and \code{n} is the sample size used (actual sample size for first measures, effective sample size for remaining ones).  The naming convention is for standard output to have \verb{(p,n)} at the end of a name or \verb{(p,n)} as subscripts in HTML and LaTeX output where \code{n} is the effective sample size and \code{p} is the number of regression parameters being adjusted for.  For indexes that are not adjusted, only \code{n} appears.
}
\description{
Generalized R^2 Measures
}
\details{
Computes various generalized R^2 measures related to the Cox-Snell R^2 for regression models fitted with maximum likelihood.  The original Cox-Snell (also due to Maddala) R^2 is labeled \code{R2} in the result.  This measure uses the raw sample size \code{n} and does not penalize for overfitting, so it can be rewarded for overfitting.  A measure adjusted for the number of fitted regression coefficients \code{p} uses the analogy to R^2 in linear models by computing \code{1 - exp(- lr / n) * (n-1)/(n-p-1)} if \code{padj=2}, which is approximately \code{1 - exp(- (lr - p) / n)}, the version used if \code{padj=1} (the default).  The latter measure is appealing because the expected value of the likelihood ratio chi-square statistic \code{lr} is \code{p} under the global null hypothesis of no predictors being associated with the response variable.  See \url{https://hbiostat.org/bib/r2.html} for more details.

It is well known that in logistic regression the Cox-Snell R^2 cannot achieve a value of 1.0 even with a perfect model, which prompted Nagelkerke to divide the R^2 measure by its maximum attainable value.  This is not necessarily the best recalibration of R^2 throughout its range.  An alternative is to use the formulas above but to replace the raw sample size \code{n} with the effective sample size, which for data with many ties can be significantly lower than the number of observations.  As used in the \code{popower()} and \code{describe()} functions, in the context of a Wilcoxon test or the proportional odds model, the effective sample size is \code{n * (1 - f)} where \code{f} is the sums of cubes of the proportion of observations at each distict value of the response variable.  To obtain R^2 measures using the effective sample size, either provide \code{ess} as a single number specifying the effective sample size, or specify a vector of frequencies of distinct Y values from which the effective sample size will be computed.  In the context of survival analysis, the single number effective sample size you may wish to specify is the number of uncensored observations.  This is exactly correct when estimating the hazard rate from a simple exponential distribution or when using the Cox PH/log-rank test.  For failure time distributions with a very high early hazard, censored observations contain enough information that the effective sample size is greater than the number of events.  See Benedetti et al, 1982.
}
\references{
Smith TJ and McKenna CM (2013): A comparison of logistic regression pseudo R^2 indices.  Multiple Linear Regression Viewpoints 39:17-26.  \url{https://www.glmj.org/archives/articles/Smith_v39n2.pdf}
Benedetti JK, et al (1982): Effective sample size for tests of censored survival data.  Biometrika 69:343--349.
Whitehead J (1993): Sample size calculations for ordered categorical data.  Stat in Med 12:2257--2271.
Mittlbock M, Schemper M (1996): Explained variation for logistic regression.  Stat in Med 15:1987-1997.
Date, S: R-squared, adjusted R-squared and pseudo R-squared. \url{https://timeseriesreasoning.com/contents/r-squared-adjusted-r-squared-pseudo-r-squared}
UCLA: What are pseudo R-squareds? \url{https://stats.oarc.ucla.edu/other/mult-pkg/faq/general/faq-what-are-pseudo-r-squareds}
Allison P (2013): What's the beset R-squared for logistic regression? \url{https://statisticalhorizons.com/r2logistic}
Menard S (2000): Coefficients of determination for multiple logistic regression analysis.  The Am Statistician 54:17-24.
}
\author{
Frank Harrell
}
