% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmClose.r
\name{rmClose}
\alias{rmClose}
\title{rmClose}
\usage{
rmClose(x, minfrac = 0.05)
}
\arguments{
\item{x}{a numeric vector with no \code{NA}s}

\item{minfrac}{minimum allowed spacing between consecutive ordered \code{x}, as a fraction of the range of \code{x}}
}
\value{
a sorted numeric vector of non-close values of \code{x}
}
\description{
Remove close values from a numeric vector that are not at the outer limtis.  This is useful for removing axis breaks that overlap when plotting.
}
\examples{
rmClose(c(1, 2, 4, 47, 48, 49, 50), minfrac=0.07)
}
\author{
Frank Harrell
}
