% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqFreq.r
\name{seqFreq}
\alias{seqFreq}
\title{seqFreq}
\usage{
seqFreq(..., labels = NULL, noneNA = FALSE)
}
\arguments{
\item{...}{any number of variables}

\item{labels}{if specified variable labels will be used in place of variable names}

\item{noneNA}{set to \code{TRUE} to not include 'none' as a level in the result}
}
\value{
\code{factor} variable with \code{obs.per.numcond} attribute
}
\description{
Find Sequential Exclusions Due to NAs
}
\details{
Finds the variable with the highest number of \code{NA}s.  From the non-\code{NA}s on that variable find the next variable from those remaining with the highest number of \code{NA}s.  Proceed in like fashion.  The resulting variable summarizes sequential exclusions in a hierarchical fashion.  See \href{https://hbiostat.org/rflow/doverview.html#sec-doverview-filter}{this} for more information.
}
\author{
Frank Harrell
}
