% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpcomp.r
\name{spikecomp}
\alias{spikecomp}
\title{spikecomp}
\usage{
spikecomp(x, count = c("table", "tabulate"), normalize = TRUE)
}
\arguments{
\item{x}{a numeric variable}

\item{count}{default is \code{"table"} to use \code{table} resulting in no zero cells; use \code{"tabulate"} to compute counts over a rigid grid, suitable for bar charts with sparklines.}

\item{normalize}{set to \code{FALSE} to not divide frequencies by maximum frequency}
}
\value{
a list with element \code{segments} which has elements \code{x}, \code{y1}, \code{y2} if \code{count='table'}, otherwise a list with elements \code{x} and \code{y}
}
\description{
Compute Elements of a Spike Histogram
}
\details{
Derives the line segment coordinates need to draw a spike histogram.  This is useful for adding elements to \code{ggplot2} plots.
}
\examples{
spikecomp(1:1000)
}
\author{
Frank Harrell
}
