% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_filter.R
\name{trim_filter}
\alias{trim_filter}
\title{From a dataframe, filter out obs with 'varname' below A & above B percent
Note that the function allows either the bottom or the ceiling to be missing.}
\usage{
trim_filter(df, varname, lowpct, uppct)
}
\description{
From a dataframe, filter out obs with 'varname' below A & above B percent
Note that the function allows either the bottom or the ceiling to be missing.
}
