% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_for_latex.R
\name{trim_for_latex}
\alias{trim_for_latex}
\title{Trim numeric digits for exporting & including in LaTex files.}
\usage{
trim_for_latex(str, pct = FALSE, dg = 4)
}
\arguments{
\item{pct}{T/F showing whether percentage outputs shoulds be exported}

\item{dg}{total number of digits to be exported}
}
\description{
Trim numeric digits for exporting & including in LaTex files.
}
