% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_vector.R
\name{trim_vector}
\alias{trim_vector}
\title{Trim a vector below A percent & above B percent.
Note that the function allows either the bottom or the ceiling to be missing.}
\usage{
trim_vector(vec, lowpct, uppct)
}
\description{
Trim a vector below A percent & above B percent.
Note that the function allows either the bottom or the ceiling to be missing.
}
