% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_l.R, R/view_m.R, R/view_s.R
\name{view_l}
\alias{view_l}
\alias{view_m}
\alias{view_s}
\title{view() with commonly used number of observations(100, 1000, 10000)}
\usage{
view_l(obj)

view_m(obj)

view_s(obj)
}
\arguments{
\item{obj}{data object}
}
\description{
view() with commonly used number of observations(100, 1000, 10000)

view() with commonly used number of observations(100, 1000, 10000)

view() with commonly used number of observations(100, 1000, 10000)
}
\section{Functions}{
\itemize{
\item \code{view_l()}: To directly "View" the resulting dataframe

\item \code{view_m()}: To directly "View" the resulting dataframe

}}
