% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlab.r
\name{vlab}
\alias{vlab}
\title{vlab}
\usage{
vlab(x, name = NULL)
}
\arguments{
\item{x}{a single variable name, unquoted}

\item{name}{optional character string to use as variable name}
}
\value{
character string
}
\description{
Easily Retrieve Text Form of Labels/Units
}
\details{
Uses the same search method as \code{hlab} returns label and units in a character string with units, if present, in brackets
}
\seealso{
\code{\link[=hlab]{hlab()}}
}
\author{
Frank Harrell
}
