\name{xtfrm.labelled}
\alias{xtfrm.labelled}
\title{
  Auxiliary Function Method for Sorting and Ranking
}
\description{
  An auxiliary function method that is a workaround for bug in the
  implementation of xtfrm handles inheritance.
}
\usage{
\method{xtfrm}{labelled}(x)
}
\arguments{
  \item{x}{
    any object of class labelled.
  }
}
\seealso{
  \code{\link{xtfrm}}
}
