\name{yearDays}
\alias{yearDays}
\alias{monthDays}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Get Number of Days in Year or Month }
\description{
  Returns the number of days in a specific year or month.
}
\usage{
yearDays(time)

monthDays(time)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{time}{
    A POSIXt or Date object describing the month or year in
    question.
  }
}
\author{ Charles Dupont }
\seealso{ \code{\link{POSIXt}}, \code{\link{Date}} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
\keyword{ chron }% __ONLY ONE__ keyword per line
