% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMarkovOrd.r
\name{estSeqMarkovOrd}
\alias{estSeqMarkovOrd}
\title{estSeqMarkovOrd}
\usage{
estSeqMarkovOrd(
  y,
  times,
  initial,
  absorb = NULL,
  intercepts,
  parameter,
  looks,
  g,
  formula,
  groupContrast = NULL,
  timecriterion = NULL,
  coxzph = FALSE,
  nsim = 1,
  progress = FALSE
)
}
\arguments{
\item{y}{vector of possible y values in order}

\item{times}{vector of measurement times}

\item{initial}{a vector of probabilities summing to 1.0 that specifies the frequency distribution of initial values to be sampled from.  The vector must have names that correspond to values of \code{y} representing non-absorbing states.}

\item{absorb}{vector of absorbing states, a subset of \code{y}.  The default is no absorbing states.  Observations are truncated when an absorbing state is simulated.}

\item{intercepts}{vector of intercepts in the proportional odds model.  There must be one fewer of these than the length of \code{y}.}

\item{parameter}{vector of true parameter (effects; group differences) values.  These are group 2:1 log odds ratios in the transition model, conditioning on the previous \code{y}.}

\item{looks}{integer vector of ID numbers at which maximum likelihood estimates and their estimated variances are computed.  For a single look specify a scalar value for \code{loops} equal to the number of subjects in the sample.}

\item{g}{a user-specified function whose first 5 arguments in order are the value of \code{y} at the previous time, the current time, the gap between the previous time and the current time, \code{group} (1 or 2), and \code{parameter} (the group effect).  The function returns the linear predictor for the proportional odds model aside from \code{intercepts}.}

\item{formula}{a formula object given to the \code{lrm()} function using variables with these name: \code{y}, \code{time}, \code{yprev}, and \code{group} (having values 1 and 2)}

\item{groupContrast}{omit this argument if \code{group} has only one regression coefficient in \code{formula}.  Otherwise provide \code{groupContrast} as a list of two lists that are passed to \code{rms::contrast.rms()} to compute the contrast of interest and its standard error.  The first list corresponds to group 2, the second to group 1, to get a 2 - 1 contrast.}

\item{timecriterion}{a function of a time-ordered vector of simulated ordinal responses \code{y} that returns a vector \code{FALSE} or \code{TRUE} values denoting whether the current \code{y} level met the condition of interest.  For example \code{estSeqMarkovOrd} will compute the first time at which \code{y >= 5} if you specify \code{timecriterion=function(y) y >= 5}.  This function is only called at the last data look for each simulated study.}

\item{coxzph}{set to \code{TRUE} if \code{timecriterion} is specified and you want to compute a statistic for testing proportional hazards at the last look of each simulated data}

\item{nsim}{number of simulations (default is 1)}

\item{progress}{set to \code{TRUE} to send current iteration number to the console}
}
\value{
a data frame with number of rows equal to the product of \code{nsim}, the length of \code{looks}, and the length of \code{parameter}.
}
\description{
Simulate Comparisons For Use in Sequential Markov Longitudinal Clinical Trial Simulations
}
\details{
Simulates sequential clinical trials of longitudinal ordinal outcomes using a first-order Markov model.  Looks are done sequentially after subject ID numbers given in the vector \code{looks} with the earliest possible look being after subject 2.  At each look, a subject's repeated records are either all used or all ignored depending on the sequent ID number.  For each true effect parameter value, simulation, and at each look, runs a function to compute the estimate of the parameter of interest along with its variance.  For each simulation, data are first simulated for the last look, and these data are sequentially revealed for earlier looks.  The user provides a function \code{g} that has extra arguments specifying the true effect of \code{parameter} the treatment \code{group} expecting treatments to be coded 1 and 2.  \code{parameter} is usually on the scale of a regression coefficient, e.g., a log odds ratio.  Fitting is done using the \code{rms} \code{lrm} function.  If \code{timecriterion} is specified, the function also, for the last data look only, computes the first time at which the criterion is satisfied for the subject.  The Cox/logrank chi-square statistic for comparing groups on the derived time variable is saved.  If \code{coxzph=TRUE}, the \code{survival} package correlation coefficient \code{rho} from the scaled partial residuals is also saved so that the user can later determine to what extent the Markov model resulted in the proportional hazards assumption being violated when analyzing on the time scale.
}
\seealso{
\code{gbayesSeqSim()}, \code{simMarkovOrd()}, \url{https://hbiostat.org/R/Hmisc/simMarkovOrd.html}
}
\author{
Frank Harrell
}
