\name{Hmisc-internal}
\title{Internal Hmisc functions}
\alias{dataDensityString}
\alias{aregTran}
\alias{as.double.Cbind}
\alias{as.numeric.Cbind}
\alias{formatSep}
\alias{as.data.frame.impute}
\alias{as.data.frame.roundN}
\alias{as.data.frame.special.miss}
\alias{as.data.frame.substi}
\alias{substi}
\alias{substi.source}
\alias{[.substi}
\alias{bpx}
\alias{convertPdate}
\alias{ddmmmyy}
\alias{expr.tree}
\alias{fillin}
\alias{formatCats}
\alias{formatCons}
\alias{formatDateTime}
\alias{formatTestStats}
\alias{format.timePOSIXt}
\alias{ftuss}
\alias{ftupwr}
\alias{get2rowHeads}
\alias{groupn}
\alias{htmlGreek}
\alias{htmlSpecial}
\alias{markupSpecs}
\alias{importConvertDateTime}
\alias{is.present}
\alias{lookupSASContents}
\alias{makeNames}
\alias{mask}
\alias{nafitted.delete}
\alias{na.include}
\alias{Names2names}
\alias{GetModelFrame}
\alias{naprint.keep}
\alias{naresid.keep}
\alias{naprint.delete}
\alias{naresid.delete}
\alias{napredict.delete}
\alias{oPar}
\alias{optionsCmds}
\alias{ordGridFun}
\alias{parGrid}
\alias{pasteFit}
\alias{plotpsummaryM}
\alias{print.substi}
\alias{print.timePOSIXt}
\alias{read.xportDataload}
\alias{readSAScsv}
\alias{rowsumFast}
\alias{sas.get.macro}
\alias{setParNro}
\alias{StatPlsmo}
\alias{stepfun.eval}
\alias{stripChart}
\alias{termsDrop}
\alias{testDateTime}
\alias{uncbind}
\alias{var.inner}
\alias{xInch}
\alias{xySortNoDupNoNA}
\alias{yInch}
\alias{zoom}
\alias{latex.responseSummary}
\alias{print.responseSummary}
\alias{responseSummary}
\alias{combine}
\alias{combine<-}
\alias{F_cidxcn}
\alias{F_cidxcp}
\alias{F_do_mchoice_match}
\alias{F_do_nstr}
\alias{F_hoeffd}
\alias{F_jacklins}
\alias{F_largrec}
\alias{F_maxempr}
\alias{F_rcorr}
\alias{F_wclosepw}
\alias{F_wclosest}
\alias{F_hlqest}

\description{Internal Hmisc functions.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
