% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.r
\name{plotlySave}
\alias{plotlySave}
\title{Save a Plotly Graphic to PNG}
\usage{
plotlySave(x, ...)
}
\arguments{
\item{x}{a plotly object, or a named list of plotly objects (excluding
objects of class \code{'plotly_hash'})}

\item{...}{additional arguments passed to \code{plotly::plotly_IMAGE}}
}
\value{
called for its side effect of writing one or more PNG files;
returns \code{invisible(NULL)}
}
\description{
\code{plotlySave} saves a plotly graphic with name \code{foo.png} where \code{foo} is
the name of the current chunk. You must have a free \code{plotly} account
from \code{plot.ly} to use this function, and you must have run
\code{Sys.setenv(plotly_username="your_plotly_username")} and
\code{Sys.setenv(plotly_api_key="your_api_key")}. The API key can be found
in one's profile settings. See
\url{http://stackoverflow.com/questions/33959635/exporting-png-files-from-plotly-in-r}.
}
\details{
The output path and base file name are taken from
\code{knitr::opts_chunk$get('fig.path')} and
\code{knitr::opts_current$get("label")} respectively. If \code{x} is a
named list (and not a \code{plotly_hash} object), one PNG file is written
per list element, with the element name appended to the chunk name.
Otherwise a single PNG file is written using just the chunk name.
}
