% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princmp.r
\name{print.princmp}
\alias{print.princmp}
\title{print.princmp}
\usage{
\method{print}{princmp}(x, which = c("none", "standardized", "original", "both"), k = x$k, ...)
}
\arguments{
\item{x}{results of \code{princmp}}

\item{which}{specifies which loadings to print, the default being \code{'none'} and other values being \code{'standardized'}, \code{'original'}, or \code{'both'}}

\item{k}{number of components to show, defaults to \code{k} specified to \code{princmp}}

\item{...}{unused}
}
\value{
nothing
}
\description{
Print Results of princmp
}
\details{
Simple print method for \code{\link[=princmp]{princmp()}}
}
\author{
Frank Harrell
}
