% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMarkovOrd.r
\name{soprobMarkovOrd}
\alias{soprobMarkovOrd}
\title{soprobMarkovOrd}
\usage{
soprobMarkovOrd(y, times, initial, absorb = NULL, intercepts, g, ...)
}
\arguments{
\item{y}{a vector of possible y values in order}

\item{times}{vector of measurement times}

\item{initial}{initial value of \code{y} (baseline state)}

\item{absorb}{vector of absorbing states, a subset of \code{y}.  The default is no absorbing states.}

\item{intercepts}{vector of intercepts in the proportional odds model, with length one less than the length of \code{y}}

\item{g}{a user-specified function of three or more arguments which in order are the value of \code{y} at the previous time, the current time, the gap between the previous time and the current time, and optional arguments such as treatment assignment (scalar).  The function returns the linear predictor for the proportional odds model aside from \code{intercepts}.  The first argument of \code{g} must be allowed to be a vector of all non-absorbing state \code{y} values.}

\item{...}{additional arguments to pass to \code{g} such as covariate settings}
}
\value{
matrix with rows corresponding to times and columns corresponding to states, with values equal to exact state occupancy probabilities
}
\description{
State Occupancy Probabilities for First-Order Markov Ordinal Model
}
\examples{
# Simulate a sample of 1000 subjects and compare proportions in
# states with exact state occupancy probabilities.  Carry absorbing
# states forward just to simplify calculation of proportions
g <- function(yprev, t, gap) yprev - ((gap - 1)/ 7) * yprev + t / 10
s <- sapply(1 : 1000, function(id)
       simMarkovOrd(1:3, c(1, 2, 3, 7, 14), 2, absorb=3,
                    intercepts=c(-1, -2), g=g, id=id, carry=TRUE), simplify=FALSE)
s <- do.call(rbind, s)
relfreq <- function(x) table(x) / length(x)
do.call(rbind, tapply(s[, 'y'], s[, 'time'], relfreq))
soprobMarkovOrd(1:3, c(1, 2, 3, 7, 14), initial=2, absorb=3,
                intercepts=c(-1, -2), g=g)
}
\author{
Frank Harrell
}
