% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpcomp.r
\name{spikecomp}
\alias{spikecomp}
\title{spikecomp}
\usage{
spikecomp(x)
}
\arguments{
\item{x}{a numeric variable}
}
\value{
a list with element \code{segments} which has elements \code{x}, \code{y1}, \code{y2}
}
\description{
Compute Elements of a Spike Histogram
}
\details{
Derives the line segment coordinates need to draw a spike histogram.  This is useful for adding elements to \code{ggplot2} plots.
}
\examples{
spikecomp(1:1000)
}
\author{
Frank Harrell
}
