\name{yearDays}
\alias{yearDays}
\alias{monthDays}
\title{ Get Number of Days in Year or Month }
\description{Returns the number of days in a specific year or month.}
\usage{
yearDays(time)

monthDays(time)
}
\arguments{
  \item{time}{
    A POSIXt or Date object describing the month or year in
    question.
  }
}
\author{ Charles Dupont }
\seealso{ \code{\link{POSIXt}}, \code{\link{Date}} }
\keyword{ utilities }
\keyword{ chron }
